log: level: fix override level resolution
CONFIG_LOG_OVERRIDE_LEVEL value was only used as a boolean to check
weather a module should have logs or not. But it's actual level was not
doing anything. It was in fact relying on the value of
CONFIG_LOG_DEFAULT_LEVEL.

In the following situation where both are defined, the final default log
level for each module was 2:

CONFIG_LOG_DEFAULT_LEVEL=2
CONFIG_LOG_OVERRIDE_LEVEL=3