drivers: watchdog: add ti TPS3851
The TPS3851 combines a precision voltage supervisor with a programmable
watchdog timer.

This driver targets the watchdog timer functionality, as the voltage
supervisor is a simple gpio read/write that does not require a
dedicated driver.

Key features:
- The watchdog timer has an output signal that can be routed to a GPIO
interrupt to trigger a user-defined callback.
- The watchdog timeouts is fixed by hardware and are defined either
by an external capacitor, or by factory-programmed default delay
settings.
- Finally, the watchdog can also be enabled/disabled via a dedicated
logic pin.

Signed-off-by: Alex Fabre <alex.fabre@rtone.fr>